requirements.txt
aiohttp==3.9.5
APScheduler==3.10.4
matplotlib==3.8.4
mplfinance==0.12.10b0
numpy==1.26.4
pandas==2.2.2
python-telegram-bot==21.4
pytz==2024.1
requests==2.32.3
TA-Lib==0.6.8
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
cryptography==42.0.8
python-dateutil==2.9.0.post0
six==1.16.0
urllib3==2.2.2
websockets==12.0  
aiofiles==24.1.0  